replacing old style dataset with new one that uses cuda streams
